+++ examples/client/tls.js
@@ -2,6 +2,6 @@
 const ddns = require('../..');
 const server = "getdnsapi.net";
 
-ddns.TLSClient.rawQuery(server,[{type: 'A', name: 'google.com', class:'IN'}]).then((result)=>{
+ddns.TLSClient.Query(server,[{type: 'A', name: 'google.com', class:'IN'}]).then((result)=>{
     console.log(result);
 })